Create updates

+++ resources/js/components/Admin/Users.vue
@@ -1,45 +1,48 @@
 <template>
     <div>
         <div class="card">
-            <div class="card-header">
-                <h5 class="card-title"></h5>
-            </div>
             <div class="card-body">
                 <div class="row">
-                    <div class="table-responsive">
+                    <div class="table-full">
                         <button @click="create.modal = true">Create</button>
                         <b-table
                         :data="users"
-                        :selected.sync="selected"
                         :paginated="true"
                         :per-page="10"
+                        width="100%"
                         :pagination-simple="false"
-                        sort-icon="arrow-up"
-                        focusable>
+                        sort-icon="arrow-up">
                             <template slot-scope="props">
-                                <b-table-column field="id" label="ID" sortable>
+                                <b-table-column field="id" label="ID" width="5%" sortable>
                                     {{ props.row.id }}
                                 </b-table-column>
                                 
-                                <b-table-column field="name" label="Name" sortable>
+                                <b-table-column field="name" label="Name" width="25%" sortable>
                                     {{ props.row.name }}
                                 </b-table-column>
                                 
-                                <b-table-column field="email" label="Email">
+                                <b-table-column field="email" label="Email" width="25%" sortable>
                                     {{ props.row.email }}
                                 </b-table-column>
                                 
-                                <b-table-column field="balance" label="Balance" sortable>
-                                    {{ props.row.balance }}
+                                <b-table-column field="balance" label="Balance" width="10%" sortable>
+                                    <b-tag type="is-info">
+                                        {{ props.row.balance }}
+                                    </b-tag>
                                 </b-table-column>
                                 
-                                <b-table-column field="is_deleted" label="Deleted" sortable>
-                                    {{ props.row.is_deleted }}
+                                <b-table-column field="is_deleted" label="Deleted" width="10%" sortable>
+                                    <b-tag v-if="props.row.is_deleted" type="is-danger">Silindi</b-tag>
+                                    <b-tag v-else type="is-success">Kayıtlı</b-tag>
                                 </b-table-column>
                                 
-                                <b-table-column label="Actions" sortable>
-                                    <span @click="editUser(user)">Edit</span>
-                                    <span @click="destroy(user)">Delete</span>
+                                <b-table-column label="Actions" width="25%">
+                                    <b-tag type="is-warning" @click.native="editUser(props.row)" title="Edit" class="action">
+                                        <b-icon icon="file-edit-outline" custom-size="mdi-18px"></b-icon>
+                                    </b-tag>
+                                    <b-tag type="is-danger" @click.native="destroyUser(props.row)" title="Delete" class="action">
+                                        <b-icon icon="close-circle" custom-size="mdi-18px"></b-icon>
+                                    </b-tag>
                                 </b-table-column>
                             </template>
                         </b-table>
@@ -50,23 +53,21 @@
 
         <!-- BEGIN::Create Modal -->
         <b-modal :active.sync="create.modal" :width="640" scroll="keep">
-            <div class="card">
-                <div class="card-content">
-                    <div class="content">
-                        <div class="form-group">
-                            <label>İsim</label>
-                            <input type="text" class="form-control" v-model="create.data.name">
-                        </div>
-                        <div class="form-group">
-                            <label>Email</label>
-                            <input type="email" class="form-control" v-model="create.data.email">
-                        </div>
-                        <div class="form-group">
-                            <label>Parola</label>
-                            <input type="password" class="form-control" v-model="create.data.password">
-                        </div>
-                        <button @click="store">Kaydet</button>
+            <div class="card-content">
+                <div class="content">
+                    <div class="form-group">
+                        <label>İsim</label>
+                        <input type="text" class="form-control" v-model="create.data.name">
                     </div>
+                    <div class="form-group">
+                        <label>Email</label>
+                        <input type="email" class="form-control" v-model="create.data.email">
+                    </div>
+                    <div class="form-group">
+                        <label>Parola</label>
+                        <input type="password" class="form-control" v-model="create.data.password">
+                    </div>
+                    <button class="btn btn-primary" @click="store">Kaydet</button>
                 </div>
             </div>
         </b-modal>
@@ -102,7 +103,6 @@
 export default {
     data(){
         return {
-            selected: {},
             create: {
                 modal: false,
                 data: {}
@@ -124,32 +124,33 @@ export default {
     methods: {
         store(){
             axios.post('/data/admin/users', this.create.data).then(response => {
-                if (response.status === 200){
-                    console.log('Başarılı')
+                if (response.status === 201){
                     axios.get('/data/admin/users/' + response.data.id).then(response => {
                         this.users.push(response.data)
                     })
                 }
                 else {
-                    console.log('Bir hata var.')
+                    console.log('There is an error!')
                 }
-                console.log(response.data)
-                this.create.modal = false;
                 this.$buefy.snackbar.open({
-                    message: response.data,
-                    type: response.status === 200 ? 'is-success' : 'is-warning',
+                    message: response.status === 201 ? "User successfully created!" : response.data,
+                    type: response.status === 201 ? 'is-success' : 'is-warning',
                     position: 'is-top',
-                    actionText: response.status === 200 ? 'OK' : 'Retry',
-                    indefinite: true,
+                    actionText: response.status === 201 ? 'OK' : 'Retry',
+                    indefinite: response.status === 201 ? false : true,
                     onAction: () => {
-                        // Retry denildiğinde yapılacaklar..
-                        this.$buefy.toast.open({
-                            message: 'Action pressed!',
-                            queue: false
-                        })
+                        if (response.status === 201){
+                            this.create.modal = false;
+                        }
                     }
                 })
             }).catch(e => {
+                this.$buefy.snackbar.open({
+                    message: e.response.data.message,
+                    type: 'is-danger',
+                    position: 'is-top',
+                    actionText: 'OK'
+                })
                 console.log(e.response)
             })
         },
@@ -190,16 +191,28 @@ export default {
             })
         },
 
-        destroy(user){
-            console.log('Silme işlemi başlatılıyor..')
-            axios.delete('/data/admin/users/' + user.id).then(response => {
-                console.log(response.data)
+        destroyUser(user){
+            this.$buefy.dialog.confirm({
+                message: 'Continue on this task?',
+                onConfirm: () => {
+                    axios.delete('/data/admin/users/' + user.id).then(response => {
+                        if (response.data === 1){
+                            this.users.splice(this.users.indexOf(user), 1)
+                            this.$buefy.toast.open('User deleted!')
+                        }
+                    })
+                }
             })
         }
     }
 }
 </script>
 
-<style>
-
+<style scoped>
+    .table-full {
+        width: 100%;
+    }
+    .action {
+        cursor: pointer;
+    }
 </style>
